Pull out Traceparent into its own class
